Fix type mismatch: use Double to match DiveTable value type
DiveTable.table stores Map<String, Map<String, Double>>;
updateAvailability parameter type corrected accordingly.

+++ android-app/app/src/main/java/com/divingdd/app/MainActivity.kt
@@ -115,7 +115,7 @@ class MainActivity : AppCompatActivity() {
         showResult("КТ = $dd")
     }
 
-    private fun updateAvailability(positionsMap: Map<String, Map<String, Float>>) {
+    private fun updateAvailability(positionsMap: Map<String, Map<String, Double>>) {
         positionBtnIds.forEach { (pos, btnId) ->
             val btn = findViewById<MaterialButton>(btnId)
             val available = positionsMap.containsKey(pos)
